Fix warning

+++ unittests/comprehensive/test_ForwardKinematics.cpp
@@ -1604,7 +1604,7 @@ Eigen::MatrixXd standardJacobian(
   Eigen::MatrixXd J = skeleton->getJacobian(node).bottomRows<3>();
 
   Eigen::MatrixXd reduced_J(3, q.size());
-  for (auto i = 0u; i < q.size(); ++i)
+  for (auto i = 0u; i < static_cast<std::size_t>(q.size()); ++i)
     reduced_J.col(i) = J.col(static_cast<int>(activeIndices[i]));
 
   return reduced_J;
